feat(hyperlink): use target ‘_blank’ by default
As the prismic content editor does not offer the option «link target» for media files (ie. PDF) its not possible for content editors to choose whether a media link should open in a new tab.

In a specific project we have the requirement to open media links in a new tab. This is a workaround while the option is missing in the content editor.